This commit adds three output types to TableDap data sets. All of the outputs are JSON messages which are specially formatted for use with the Google Visualization client library (Google Charts).
The first, .dataTable, makes no assumptions about the contents and does nothing to the order of the variables. The columns
are listed in the order they appear in the URL. This is good for scatter plots of say salinity vs. temperature.

The second, .dataTableMap, requires the presence of a latitude and a longitude variable and changes the order of the columns
so latitude is first, followed by longitude as required for using the Google GeoCharts.

The final type is, .dataTableTimeSeries. To make a timeseries plot using the AnnotationChart the data must include a datetime
column and that column must be listed first. The columns will be re-ordered so time is listed first.

If any of the required columns are missing from a request, the TableWriter will throw an exception. The messages for the
exceptions were not externalized.