Optimize maxStringLength
Because StringHolders store the strings as char array, using that directly instead of converting to a string saves a lot of time (and heap allocations).

Prior to this chanve Table.testReadInvalidCRA() ran in 117,394 ms with 72,509 ms being spent in StringArray.maxStringLength

With this change total time is 38,999 ms with StringArray.maxStringLength at 41.2 ms. There were also speed ups outside of maxStringLength which I believe is due to reducing GC.


Former-commit-id: ff2da20b4048bbb99bcef0d88b8b1d0d4d31b7ca [formerly f50a97de86d02d5101db2d6bbf746e17317c7cfa] [formerly 95c8bde96a3027a4dc2ccdab5c9576c25b76fde5 [formerly 299795bc7fa516a3a8029c263057bbffeef319f3]]
Former-commit-id: c5c0e950ca2a3425dca7678946823932bd9f2abc [formerly 0ca29f6b1717ff07262c6c0ad5c3d50e0150e0fd]
Former-commit-id: 6c0ce899fed9ec8bd5d014d13a5fddddbb3064f9
Former-commit-id: 49233efad69d60b2c8eb5b27cec0fcef12ec5244