Update to new Google authentication api
Also have login automatically redirect to the loginUrl instead of requiring the user to click on an additional button. This is a better user experience, but more importantly the new Google sign in button never changes state to say it's logged in so the user no longer had an indication that the Google sign in was successful.


Former-commit-id: c3b668b049603e9a28b61d880f710f57c8effc9e [formerly 3df33bdb5922ced32c755be14e77f82c2a116f0c] [formerly d669b4ddbc4d34a7d16d999cd0eefe33ddd34f48 [formerly 0fd90b55a4448e11748893fe5ca8ea959a9ac501]]
Former-commit-id: cc2393db7daac96a606a228a02c5f9c90b53cb92 [formerly 08df3f83932a0f61a320c25c026fadad1d7329d0]
Former-commit-id: 47fb1dec9d2cdfe3666e61404f1bb66f9b101e47
Former-commit-id: 6bc62dc3185869d18a521b1afa2c06a38c78c96e