Update to new Google authentication api
Also have login automatically redirect to the loginUrl instead of requiring the user to click on an additional button. This is a better user experience, but more importantly the new Google sign in button never changes state to say it's logged in so the user no longer had an indication that the Google sign in was successful.