Add GitHub Actions maven build and test for main branch (#208)
* Add GitHub Actions maven build and test for main branch

* Adds GitHub Action in `.github/workflows/maven.yml` to build and test on pushes or pull_requests to `main`
* Moves test data from `src/test/resources` to `test-data` and adds `test-data` as a `maven-surefire-plugin` `additionalClasspathElement`
* Removes `download-maven-plugin` cacheDirectory overrides, moving cache directory to default `.m2/respository/.cache/maven-download-plugin`
* Fixes `TopLevelHandlerTests` which alters the value of `EDStatic.angularDegreeUnitsSet`, affecting subsequent tests (test order is indeterminate)
* Updates `development/docker/Dockerfile` (adds missing .mvn dir needed for `errorprone`, skips git-code-format hook installtion, and updates Java version to 21)

The test data move from `src/test/resources` to `test-data` was needed because many of the test cases require a specific data file to be
the last file (latest mtime) in the matching set. `maven-resources-plugin` does not preserve mtime when copying test resources into
`target/test-classes`, which was leading to files in test datasets having exactly the same mtime (the time at which `maven-resources-plugin`
copied the files), and an indeterminate file was being selected as "latest". By using the test data in place via `additionalClasspathElement`,
we preserve the mtimes and also avoid copying huge amounts of data into `target/test-classes` (win win).

Note: `gha-test` branch reference in `.github/workflows/maven.yml` can be deleted just before or after merge. It's just there to demonstrate
working build/test.

* Add enforcer plugin check for test data in src/test/resources

* Add test-data to the path for failsafe tests.
Move the BPD for the test setup to under target so that it is cleared when running mvn clean.

* GHA: include integration tests (mvn verify)

---------

Co-authored-by: Chris John <chris.john@noaa.gov>