docker-compose: wait for erddap container startup to start logs container
As reported in #271, the `logs` container starts at the same time
as the `erddap` container, and during a first time start up complains
that the `/erddapData/logs/log.txt` file doesn't exist. Once the
`erddap` container starts up the `log.txt` file will be created and
due to the `tail -F` retry behavior it will be detected and tailed.

This change makes the `logs` container start up wait for the `erddap`
container to be up and healthy before tailing the log file, avoiding
the harmless but confusing warning message.

Fixes #271.