Support useHeadersForUrl feature flag (use Host header when building urls)
Add support for a useHeadersForUrl feature flag, which changes
EDStatic methods erddapUrl() and erddapHttpsUrl() to use request
`Host` header when building urls if useHeadersForUrl is true
and the request is available/not null.

Functionally, when this flag is enabled ERDDAP can accessed
on any address which resolves to it. For example an ERDDAP
available at https://erddap.officialdomain.org/erddap,
http://server.localdomain:8080/erddap, and http://localhost:8080/erddap
will generate URLs consistent with the accessed hostname.

This change makes the assumption that most URLs written in response to
user interaction (e.g. loading a page in a browser) should
respect the Host header in the request, but some URLs
(example RSS feeds) should continue to always use the
erddapPreferredUrl, which is constructed from baseUrl and baseHttpsUrl
uses baseHttpsUrl if available, or baseUrl if not). If the
header driven URL constructing approach becomes the default in the
future, we could imagine only having to supply a single
ERDDAP preferred URL config to use when constructing URLs outside
of a request context (generated emails, etc) instead of separate
baseUrl and baseHttpsUrl.

One corner case is the behavior when a user accesses a login page
over http. When useHeadersForUrl is enabled and EDStatic.erddapHttpsUrl
is called with an http request, the resulting URL prefix is upgraded to
https with the request provided Host header as long as there is not a
port in the request Host (i.e. a request for http://erddap.domain.org
would result in a https://erddap.domain.org base URL, but a request for
http://erddap.domain.org:8080 would fall back to returning
the erddapHttpsUrl instead since https://erddap.domain.org:8080 is
nonsensical).

Absolute URLs are constructed from the request Host header instead of
path-only ("/erddap/index.html") or relative ("../status.html") URLs.
Returning path-only URLs would be technically simple, and I originally
included an option in EDStatic.erddapUrl() to generate them, but removed
it for simplicity. EDStatic.erddapUrl() is called a vast number of times,
and the risk of bug introduction by generation of path-only or relative URLs
(non-interactive contexts etc) seemed to justify keeping the URLs absolute,
at least for this change. As such, this change doesn't implement relative
URLs but does address many of the concerns in #96.

General expectations of behavior:

If useHeadersForUrl is false, erddapUrl or erddapHttpsUrl are used (legacy behavior).

If the request is null, erddapUrl or erddapHttpsUrl are used (legacy behavior).

If useHeadersForUrl is true and request is not null, scheme and host header from
request are used.

If useHeadersForUrl true, request not null, and request scheme is http,
EDStatic.erddapHttpsUrl() uses https scheme with request host as long as request host does not
contain a port (i.e. http url are upgraded to https with the same host).
If host does contain a port, erddapHttpsUrl is returned instead.