Release Docker images
This builds on #331 to enable building of Docker images on tagging 
Releases.

Using the Docker Metadata Action, it generates a different set of tags 
as needed based on the event (push commit to main, push tag to main).

When a commit is pushed to main, it should generate `alpha-latest` and 
date tagged images. When a (semver compatible) tag is pushed, it should 
generate an image that has that tag.